requirements.txt
streamlit
torch
transformers
diffusers
Pillow
opencv-python
numpy
matplotlib
pandas 